refactor(ai/test): split domain_planner_orchestrator_test into 3 files
Replaces the monolithic 760-line `domain_planner_orchestrator_test.dart`
with focused files aligned to planner surfaces under `lib/src/planning/`:

- `domain_planner_orchestrator_move_orders_test.dart` — default API move
  and army-move-from-game-setup cases.
- `domain_planner_orchestrator_army_move_fake_api_test.dart` — fake API
  army move injection.
- `domain_planner_orchestrator_cross_planner_test.dart` — economy, naval,
  research, diplomacy, and build-picker branches.

Test bodies unchanged; only file boundaries and imports adjusted.

Refs #2289
